feat(component): add product page styling and cart notification
